Move tests to abstract
